refactor(KERN-1560): split helpers.ts into focused modules
- aggregate-detection.ts: fetchAggregateFunctions, isAggregateExpr,
  isNestedAggregateExpr, isWindowExpr, isStarExpr
- naming.ts: exprToName, generateAggregateName, deduplicateName,
  inferTypeFromExpr
- ast-utils.ts: getConstantValue, getColumnName, getQualifiedColumnRef,
  extractTableName, matchMeasureFromExpr, hasRecursiveCteInMap, etc.
- helpers.ts: barrel re-export (all existing imports continue to work)

work-item: KERN-1560
